requirements.txt
streamlit~=1.47.1
google-genai 
protobuf~=6.31.1
openai
anthropic
groq